chatbot/policy_api.py
import os
import logging
import requests
import json

logger = logging.getLogger(__name__)

class PolicyAPI:
    """
    API client for fetching abortion policy information by state
    """
    def __init__(self):
        """Initialize the Policy API client"""
        logger.info("Initializing Policy API")
        try:
            # The API endpoint would typically be set in environment variables
            self.api_endpoint = os.environ.get("POLICY_API_ENDPOINT", "https://api.example.org/abortion-policy")
            self.api_key = os.environ.get("POLICY_API_KEY", "default_key")
            
            logger.info("Policy API initialized successfully")
        except Exception as e:
            logger.error(f"Error initializing Policy API: {str(e)}", exc_info=True)
            raise
    
    def get_policy_response(self, question):
        """
        Get policy information based on the user's question
        
        Args:
            question (str): User's policy-related question
        
        Returns:
            str: Response with policy information
        """
        try:
            # Extract state information from the question if present
            state = self._extract_state_from_question(question)
            
            if not state:
                return self._get_general_policy_response(question)
            
            # Make API request for state-specific policy
            return self._get_state_policy(state, question)
            
        except Exception as e:
            logger.error(f"Error getting policy response: {str(e)}", exc_info=True)
            return "I'm sorry, I'm having trouble retrieving the latest policy information. For the most up-to-date and accurate information, please consult the Planned Parenthood website or contact a healthcare provider directly."
    
    def _extract_state_from_question(self, question):
        """
        Extract state name from the question
        
        Args:
            question (str): User's question
        
        Returns:
            str or None: State name if found, None otherwise
        """
        # List of US states
        states = [
            "Alabama", "Alaska", "Arizona", "Arkansas", "California", "Colorado", "Connecticut", 
            "Delaware", "Florida", "Georgia", "Hawaii", "Idaho", "Illinois", "Indiana", "Iowa", 
            "Kansas", "Kentucky", "Louisiana", "Maine", "Maryland", "Massachusetts", "Michigan", 
            "Minnesota", "Mississippi", "Missouri", "Montana", "Nebraska", "Nevada", "New Hampshire", 
            "New Jersey", "New Mexico", "New York", "North Carolina", "North Dakota", "Ohio", 
            "Oklahoma", "Oregon", "Pennsylvania", "Rhode Island", "South Carolina", "South Dakota", 
            "Tennessee", "Texas", "Utah", "Vermont", "Virginia", "Washington", "West Virginia", 
            "Wisconsin", "Wyoming"
        ]
        
        question_lower = question.lower()
        
        # Check for state names in the question
        for state in states:
            if state.lower() in question_lower:
                return state
        
        return None
    
    def _get_general_policy_response(self, question):
        """
        Get a general policy response when no specific state is mentioned
        
        Args:
            question (str): User's policy-related question
        
        Returns:
            str: General policy information
        """
        return """
        Abortion laws and policies vary considerably by state in the United States. Some states have strict restrictions or bans, while others protect access to abortion.

        For the most accurate and up-to-date information about abortion laws in a specific state, please:
        1. Mention the state you're asking about in your question
        2. Visit the Planned Parenthood website (plannedparenthood.org)
        3. Contact a healthcare provider in your area
        4. Call the Planned Parenthood helpline at 1-800-230-PLAN (7526)

        Would you like to know about abortion laws in a specific state?
        """
    
    def _get_state_policy(self, state, question):
        """
        Get state-specific policy information
        
        Args:
            state (str): State name
            question (str): Original question
        
        Returns:
            str: State-specific policy information
        """
        try:
            # In a real implementation, this would make an API call
            # For this implementation, we'll simulate the API response
            # with placeholder text that would be replaced by real data
            
            logger.debug(f"Fetching policy information for state: {state}")
            
            # Simulate API call delay and response
            # In a real implementation, you would use:
            # response = requests.get(
            #     f"{self.api_endpoint}/state/{state}",
            #     headers={"Authorization": f"Bearer {self.api_key}"}
            # )
            # policy_data = response.json()
            
            # Placeholder response
            return f"""
            Based on the most recent information available for {state}, here is what I can tell you about abortion policies:
            
            This would be replaced with actual policy information from the API response. The information would include details about:
            
            - Legal status of abortion
            - Gestational limits
            - Waiting periods
            - Parental consent requirements
            - Insurance coverage
            - Facility requirements
            - Available services
            
            Please note that abortion laws can change, so for the most up-to-date information, I recommend:
            
            1. Contacting Planned Parenthood directly
            2. Visiting the Planned Parenthood website
            3. Consulting with a healthcare provider in {state}
            
            Would you like me to provide information about resources available in {state}?
            """
            
        except Exception as e:
            logger.error(f"Error getting state policy for {state}: {str(e)}", exc_info=True)
            return f"I'm sorry, I'm having trouble retrieving the latest policy information for {state}. For the most accurate information, please consult the Planned Parenthood website or contact a healthcare provider directly."
